Features/bo backend corregir tipos de propiedades de las clases

+++ src/main/java/com/ccsw/bidoffice/filetype/model/FileTypeEntity.java
@@ -20,7 +20,7 @@ public class FileTypeEntity {
     private String name;
 
     @Column(name = "priority", nullable = false)
-    private Long priority;
+    private Integer priority;
 
     public Long getId() {
         return id;
@@ -38,11 +38,12 @@ public class FileTypeEntity {
         this.name = name;
     }
 
-    public Long getPriority() {
+    public Integer getPriority() {
         return priority;
     }
 
-    public void setPriority(Long priority) {
+    public void setPriority(Integer priority) {
         this.priority = priority;
     }
+
 }
